requirements.txt
asgiref==3.5.2
attrs==21.4.0
celery==5.2.3
Django==3.2.7
django-celery-results==2.2.0
django-cors-headers==3.12.0
django-environ==0.8.1
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-errors-formatter==1.0.1
drf-spectacular==0.22.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.5.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
redis==3.5.3
sqlparse==0.4.2
uritemplate==4.1.1
